fixed Modal in TTT
