+++ front-end/app/tic-tac-toe/hook/useSocketEvent.tsx
@@ -45,15 +45,15 @@ export const useSocketEvent = () => {
 						number: clients,
 						symbol: clients === 0 ? 'X' : clients === 1 ? 'O' : '🔳',
 				}
-				newPlayerData.players[2] = {
-					name: "KEK",
-					color: 0xff0000,
-					number: 1,
-					symbol: '🔳'
-				}
+				// newPlayerData.players[2] = {
+				// 	name: "KEK",
+				// 	color: 0xff0000,
+				// 	number: 1,
+				// 	symbol: '🔳'
+				// }
 				newPlayerData.client = clients
 				updatePlayerState( newPlayerData );
-				updateGameState({...gameState, bot: clients === 0 ? true : false});
+				updateGameState({...gameState, bot: false});
 			}
 		};
