+++ front-end/app/tic-tac-toe/hook/useUI.tsx
@@ -1,9 +1,11 @@
 import { useSound } from "@/components/Sound";
-import { useState } from "react";
+import { useEffect, useState } from "react";
+import { useGameState } from "./useGameState";
 
 export const useUI = () => {
 	const [showModal, setShowModal] = useState(false);
 	const soundEngine = useSound();
+	const { gameState } = useGameState();
 
 	const closeModal = () => {
 		setShowModal(false);
@@ -15,6 +17,21 @@ export const useUI = () => {
 		//losing2();
 		setShowModal(true);
 	}
+
+
+	// Opens the EndModal after a delay if the 'gameOver' state is true.
+	useEffect(() => {
+		if (gameState.gameOver) {
+			const delay = 2000;
+			const modalTimeout = setTimeout(() => {
+				openModal();
+			}, delay);
+
+			return (() => {
+				clearTimeout(modalTimeout)
+			});
+		}
+	}, [gameState.gameOver]);
 
 	return {
 		closeModal,
